pull out prop string
